finish logic and styling

+++ apps/modernization-ui/src/apps/page-builder/pages/BusinessRulesLibrary/Edit/EditBusinessRules.tsx
@@ -1,12 +1,14 @@
-import { Button, ButtonGroup, Form, Grid, Icon } from '@trussworks/react-uswds';
+import { Button, ButtonGroup, Form, Grid, Icon, ModalRef, ModalToggleButton } from '@trussworks/react-uswds';
 import { PagesBreadcrumb } from 'apps/page-builder/components/PagesBreadcrumb/PagesBreadcrumb';
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useContext, useEffect, useRef, useState } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 import { useNavigate, useParams } from 'react-router-dom';
 import { UserContext } from 'user';
 import './EditBusinessRules.scss';
 // import { EditBusinessRulesFields } from './EditBusinessRulesFields';
 import { PageRuleControllerService, ViewRuleResponse } from '../../../generated';
+import DeleteBusinessRuleWarningModal from './DeleteBusinessRuleWarningModal';
+import { useAlert } from 'alert';
 
 export type FormValues = {
     ruleFunction: string;
@@ -31,15 +33,16 @@ export const EditBusinessRules = () => {
         mode: 'onChange'
     });
     const token = `Bearer ${state.getToken()}`;
+    const deleteWarningModalModal = useRef<ModalRef>(null);
+    const { showAlert } = useAlert();
 
     useEffect(() => {
         if (!ruleId) return;
         PageRuleControllerService.viewRuleResponseUsingGet({
             authorization: token,
-            ruleId: Number(ruleId),
-            page: Number(pageId)
+            ruleId: Number(ruleId)
         }).then((resp: ViewRuleResponse) => {
-            form.setValue('anySourceValue', resp?.anySourceValue!);
+            // form.setValue('anySourceValue', resp?.!);
             form.setValue('ruleDescription', resp?.ruleDescription!);
             form.setValue('ruleFunction', resp?.ruleFunction!);
             form.setValue('comparator', resp?.comparator!);
@@ -84,6 +87,8 @@ export const EditBusinessRules = () => {
                 ruleId: Number(ruleId),
                 request
             }).then((resp: any) => {
+                console.log(resp);
+                console.log(resp);
                 // showAlert({ type: 'success', header: 'updated', message: resp.message });
             });
         }
@@ -94,14 +99,23 @@ export const EditBusinessRules = () => {
         navigate(-1);
     };
 
-    const handleDeleteRule = () => {
-        PageRuleControllerService.deletePageRuleUsingDelete({
-            authorization: token,
-            ruleId: Number(ruleId)
-        }).then((resp: any) => {
-            handleCancel();
-            // showAlert({ type: 'success', header: 'Deleted', message: resp.message });
-        });
+    const handleDeleteRule = async () => {
+        console.log('delete rule', { pageId, ruleId });
+        if (pageId) {
+            try {
+                const response = await PageRuleControllerService.deletePageRuleUsingDelete({
+                    authorization: token,
+                    pageId: pageId ?? '',
+                    ruleId: Number(ruleId)
+                });
+                // show alert
+                console.log('response', response);
+                showAlert({ type: 'success', header: 'deleted', message: response.message });
+            } catch (error: any) {
+                showAlert({ type: 'error', header: 'error', message: error.message });
+                console.error('Error', error);
+            }
+        }
     };
 
     const fieldTypeTab = [
@@ -117,65 +131,76 @@ export const EditBusinessRules = () => {
     const ruleFunction = form.watch('ruleFunction');
 
     return (
-        <div className="edit-rules">
-            <div className="breadcrumb-wrap">
-                <PagesBreadcrumb currentPage={`${title} business rules`} />
-            </div>
-            <Form onSubmit={onSubmit}>
-                <div className="edit-rules__form">
-                    <div className="edit-rules__content">
-                        <h2>{`${title} business rules`}</h2>
-                        <Grid row className="inline-field">
-                            <Grid col={3}>
-                                <label className="input-label">Function</label>
+        <>
+            <DeleteBusinessRuleWarningModal
+                deleteWarningModalModalRef={deleteWarningModalModal}
+                onDeleteRule={handleDeleteRule}
+            />
+            <div className="edit-rules">
+                <div className="breadcrumb-wrap">
+                    <PagesBreadcrumb currentPage={`${title} business rules`} />
+                </div>
+                <Form onSubmit={onSubmit}>
+                    <div className="edit-rules__form">
+                        <div className="edit-rules__content">
+                            <h2>{`${title} business rules`}</h2>
+                            <Grid row className="inline-field">
+                                <Grid col={3}>
+                                    <label className="input-label">Function</label>
+                                </Grid>
+                                <Grid col={8}>
+                                    {ruleId ? (
+                                        <label>{ruleFunction}</label>
+                                    ) : (
+                                        <ButtonGroup type="segmented">
+                                            {fieldTypeTab.map((field, index) => (
+                                                <Button
+                                                    key={index}
+                                                    type="button"
+                                                    outline={field.name !== selectedFieldType}
+                                                    onClick={() => {
+                                                        setSelectedFieldType(field.name);
+                                                        form.setValue('ruleFunction', field.name);
+                                                    }}>
+                                                    {field.name}
+                                                </Button>
+                                            ))}
+                                        </ButtonGroup>
+                                    )}
+                                </Grid>
                             </Grid>
-                            <Grid col={8}>
-                                {ruleId ? (
-                                    <label>{ruleFunction}</label>
-                                ) : (
-                                    <ButtonGroup type="segmented">
-                                        {fieldTypeTab.map((field, index) => (
-                                            <Button
-                                                key={index}
-                                                type="button"
-                                                outline={field.name !== selectedFieldType}
-                                                onClick={() => {
-                                                    setSelectedFieldType(field.name);
-                                                    form.setValue('ruleFunction', field.name);
-                                                }}>
-                                                {field.name}
-                                            </Button>
-                                        ))}
-                                    </ButtonGroup>
-                                )}
-                            </Grid>
-                        </Grid>
-                        {selectedFieldType == '' ? (
-                            <></>
+                            {selectedFieldType == '' ? (
+                                <></>
+                            ) : (
+                                <FormProvider {...form}>{/* <EditBusinessRulesFields /> */}</FormProvider>
+                            )}
+                        </div>
+                    </div>
+                    <div className="edit-rules__buttons">
+                        {ruleId ? (
+                            <ModalToggleButton
+                                opener
+                                modalRef={deleteWarningModalModal}
+                                type="button"
+                                className="delete-btn"
+                                unstyled>
+                                <Icon.Delete size={3} className="margin-right-2px" />
+                                <span> Delete</span>
+                            </ModalToggleButton>
                         ) : (
-                            <FormProvider {...form}>{/* <EditBusinessRulesFields /> */}</FormProvider>
+                            <div />
                         )}
+                        <div>
+                            <Button type="button" outline onClick={handleCancel}>
+                                Cancel
+                            </Button>
+                            <Button type="submit" className="lbr" disabled={!form.formState.isValid}>
+                                Add to Library
+                            </Button>
+                        </div>
                     </div>
-                </div>
-                <div className="edit-rules__buttons">
-                    {ruleId ? (
-                        <Button type="button" className="delete-btn" unstyled onClick={handleDeleteRule}>
-                            <Icon.Delete size={3} className="margin-right-2px" />
-                            <span> Delete</span>
-                        </Button>
-                    ) : (
-                        <div />
-                    )}
-                    <div>
-                        <Button type="button" outline onClick={handleCancel}>
-                            Cancel
-                        </Button>
-                        <Button type="submit" className="lbr" disabled={!form.formState.isValid}>
-                            Add to Library
-                        </Button>
-                    </div>
-                </div>
-            </Form>
-        </div>
+                </Form>
+            </div>
+        </>
     );
 };
